Refactor generation of integers
Creating a separate buffer and copying it afterwards is a waste of
memory and CPU cycles. Insert the result directly in the buffer provided
by the user. This has the advantage to let us get rid of the ugly hack
to determine the size of the buffer for the integer.

While we are at it, inline the functions - both were used only once.