Merge arrays with objects
This is in preparation to be more conforming with RFC 8259. JSON text
does not need to be an object - use a generic type. Add .stype member to
define which structured type (object, array) to create.

'.type' describes the type of '.value', rename it to '.vtype' to make
this more clear.